Require files from the .cache directory so user can run es/commonjs (#30338)
* Require files from the .cache directory so user can run es/commonjs

* Panic on build if we get an error

* Add some logging via reporter

* logging to func

* Update packages/gatsby-plugin-functions/src/gatsby-node.ts


* Function name does not include extension

* Print a relative path to the functions directory

+++ packages/gatsby-plugin-functions/src/gatsby-node.ts
@@ -21,7 +21,7 @@ export async function onPreBootstrap(
     path: functionsDirectoryPath = DEFAULT_FUNCTIONS_DIRECTORY_PATH,
   }: PluginOptions
 ): Promise<void> {
-  const activity = reporter.activityTimer(`building functions`)
+  const activity = reporter.activityTimer(`Compiling Gatsby Functions`)
   activity.start()
 
   const functionsDirectory = path.resolve(
@@ -34,23 +34,19 @@ export async function onPreBootstrap(
   // Get initial list of files
   const files = await glob(functionsGlob, { cwd: functionsDirectory })
 
-  // console.log({ files })
-
-  const knownFunctions = new Set(files)
-
-  // console.log({ knownFunctions })
-
-  // if (![`develop`, `build-javascript`].includes(stage)) {
-  //   return Promise.resolve()
-  // }
+  if (files?.length === 0) {
+    reporter.warn(
+      `No functions found in directory: ${path.relative(
+        process.cwd(),
+        functionsDirectory
+      )}`
+    )
+    return
+  }
 
   await fs.ensureDir(path.join(process.cwd(), `.cache`, `functions`))
 
   await fs.emptyDir(path.join(process.cwd(), `.cache`, `functions`))
-
-  // console.log({
-  //   path: path.join(functionsDirectory, `month.js`),
-  // })
 
   const gatsbyVarObject = Object.keys(process.env).reduce((acc, key) => {
     if (key.match(/^GATSBY_/)) {
@@ -69,51 +65,56 @@ export async function onPreBootstrap(
     }
   )
 
-  await Promise.all(
-    files.map(file => {
-      const config = {
-        entry: path.join(functionsDirectory, file),
-        output: {
-          path: path.join(process.cwd(), `.cache`, `functions`),
-          filename: file.replace(`.ts`, `.js`),
-          libraryTarget: `commonjs2`,
-        },
-        target: `node`,
-        // library: "yourLibName",
+  try {
+    await Promise.all(
+      files.map(file => {
+        const config = {
+          entry: path.join(functionsDirectory, file),
+          output: {
+            path: path.join(process.cwd(), `.cache`, `functions`),
+            filename: file.replace(`.ts`, `.js`),
+            libraryTarget: `commonjs2`,
+          },
+          target: `node`,
+          // library: "yourLibName",
 
-        mode: `production`,
-        module: {
-          rules: [
-            {
-              test: [/.js$/, /.ts$/],
-              exclude: /node_modules/,
-              loader: `babel-loader`,
-            },
-          ],
-        },
-        plugins: [new webpack.DefinePlugin(varObject)],
-        // devtool: `source-map`,
-      }
+          mode: `production`,
+          module: {
+            rules: [
+              {
+                test: [/.js$/, /.ts$/],
+                exclude: /node_modules/,
+                loader: `babel-loader`,
+              },
+            ],
+          },
+          plugins: [new webpack.DefinePlugin(varObject)],
+          // devtool: `source-map`,
+        }
 
-      return new Promise((resolve, reject) =>
-        // if (stage === `develop`) {
-        //   webpack(config).watch({}, () => {})
+        return new Promise((resolve, reject) =>
+          // if (stage === `develop`) {
+          //   webpack(config).watch({}, () => {})
 
-        //   return resolve()
-        // }
+          //   return resolve()
+          // }
 
-        webpack(config).run((err, stats) => {
-          console.log({
-            warnings: stats.compilation.warnings,
+          webpack(config).run((err, stats) => {
+            if (stats?.compilation?.warnings?.length > 0) {
+              reporter.warn(stats.compilation.warnings)
+            }
+
+            if (err) return reject(err)
+            const errors = stats.compilation.errors || []
+            if (errors.length > 0) return reject(stats.compilation.errors)
+            return resolve()
           })
-          if (err) return reject(err)
-          const errors = stats.compilation.errors || []
-          if (errors.length > 0) return reject(stats.compilation.errors)
-          return resolve()
-        })
-      )
-    })
-  )
+        )
+      })
+    )
+  } catch (e) {
+    activity.panic(`Failed to compile Gatsby Functions.`, e)
+  }
 
   activity.end()
 }
@@ -131,7 +132,7 @@ export async function onCreateDevServer(
   )
   const files = await glob(functionsGlob, { cwd: functionsDirectory })
 
-  // console.log({ files })
+  reporter.verbose(`Attaching functions to development server`)
 
   const knownFunctions = new Map(
     files.map(file => [
@@ -139,7 +140,6 @@ export async function onCreateDevServer(
       file,
     ])
   )
-  // console.log(app, functionsDirectoryPath)
 
   app.use(
     `/api/:functionName`,
@@ -152,12 +152,27 @@ export async function onCreateDevServer(
       const { functionName } = req.params
 
       if (knownFunctions.has(functionName)) {
-        const fn = require(path.join(
-          functionsDirectory,
-          knownFunctions.get(functionName) as string
-        ))
+        const activity = reporter.activityTimer(
+          `Executing function ${functionName}`
+        )
+        activity.start()
+        const compiledFunctionsDir = path.join(
+          process.cwd(),
+          `.cache`,
+          `functions`
+        )
+        const funcNameToJs = knownFunctions.get(functionName) as string
 
-        fn(req, res)
+        try {
+          const fn = require(path.join(compiledFunctionsDir, funcNameToJs))
+
+          const fnToExecute = (fn && fn.default) || fn
+
+          fnToExecute(req, res)
+        } catch (e) {
+          reporter.error(e)
+        }
+        activity.end()
       } else {
         next()
       }
